% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.fctnum.R
\name{as.fctnum}
\alias{as.fctnum}
\title{Convert from factor into numeric vector}
\usage{
as.fctnum(x)
}
\arguments{
\item{x}{a vector of factor whose elements are numeric}
}
\value{
a vector as numeric
}
\description{
\code{as.fctnum} Convert the class of vector from factor into numeric
}
\examples{
# as.fctnum(y)
}
