% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsp_palette.R
\name{dsp_palette}
\alias{dsp_palette}
\title{Display color palette indicated by user}
\usage{
dsp_palette(color, size = length(color))
}
\arguments{
\item{color}{a character vector including colors (color name or hex)}

\item{size}{how many colors are displayed (default is the same length as color vector)}
}
\value{
display colors rectangle
}
\description{
\code{dsp_palette} Display the colors user indicated as the discrete rectangle
}
\examples{
# dsp_palette(c("red", "blue", "#444444"))
}
