% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib.all.R
\name{lib.all}
\alias{lib.all}
\title{Load multiple packages at once}
\usage{
lib.all(..., suppress = TRUE)
}
\arguments{
\item{...}{package names, either single string or multiple vector is OK.}

\item{suppress}{logical whether suppress packages startup messages (default is TRUE).}
}
\value{
execution of loading libraries
}
\description{
\code{lib.all} Load all indicated packages
}
\examples{
pkgs_all <- c("tidyr", "dplyr", "ggplot2")
lib.all(pkgs_all, suppress = FALSE)
}
