% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg.not.installed.R
\name{pkg.not.installed}
\alias{pkg.not.installed}
\title{Check uninstalled packages}
\usage{
pkg.not.installed(...)
}
\arguments{
\item{...}{package names, either single string or multiple vector is OK.}
}
\value{
character strings
}
\description{
\code{pkg.not.installed} Check uninstalled packages
}
\examples{
pkgs_all <- c("tidyr", "dplyr", "ggplot2")
pkg.not.installed(pkgs_all)
}
