% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgs.install.R
\name{pkgs.install}
\alias{pkgs.install}
\title{Multiple packages installation at once}
\usage{
pkgs.install(...)
}
\arguments{
\item{...}{package names, either single string or multiple vector is OK.}
}
\value{
execution of packages install
}
\description{
\code{pkgs.install} Install indicated packages at once
}
\examples{
pkgs_all <- c("tidyr", "dplyr", "ggplot2")
pkgs.install(pkgs_all)
}
